refactor: 음식점 객체를 관리하는 함수 위치를 변경
