+++ src/components/RestaurantList/handleRestaurantList.ts
@@ -1,8 +1,4 @@
 import RestaurantList from ".";
-import { restaurants } from "../../domain/restaurants";
-import defaultRestaurants from "../../tools/defaultRestaurants";
-import Storage from "../../tools/Storage";
-import IRestaurant from "../../type/IRestaurant";
 import BottomSheet from "../BottomSheet";
 
 export const renderRestaurantList = () => {
@@ -27,20 +23,4 @@ export const onClickRestaurantItem = (id: string) => {
   if (bottomSheet instanceof BottomSheet) {
     bottomSheet.open(`<restaurant-view restaurant-id=${id}></restaurant-view>`);
   }
-};
-
-// 다른 폴더로 이동 예정
-export const addRestaurant = (newRestaurant: IRestaurant) => {
-  restaurants.state.restaurants = [
-    ...restaurants.state.restaurants,
-    newRestaurant,
-  ];
-  Storage.saveRestaurants(restaurants.state.restaurants);
-};
-
-// 다른 폴더로 이동 예정
-export const restoreRestaurants = () => {
-  const restoredRestaurants = Storage.loadRestaurants();
-  restaurants.state.restaurants =
-    restoredRestaurants.length > 0 ? restoredRestaurants : defaultRestaurants;
 };
